Move export functionality to AnomalyModule
