+++ src/anomalib/deploy/__init__.py
@@ -3,15 +3,10 @@
 # Copyright (C) 2022-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-from .export import ExportType, export_to_onnx, export_to_openvino, export_to_torch
 from .inferencers import Inferencer, OpenVINOInferencer, TorchInferencer
 
 __all__ = [
-    "ExportType",
     "Inferencer",
     "OpenVINOInferencer",
     "TorchInferencer",
-    "export_to_onnx",
-    "export_to_openvino",
-    "export_to_torch",
 ]
