+++ src/anomalib/models/components/base/__init__.py
@@ -3,9 +3,9 @@
 # Copyright (C) 2022-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-from .anomaly_module import AnomalyModule
+from .anomaly_module import AnomalyModule, ExportType
 from .buffer_list import BufferListMixin
 from .dynamic_module import DynamicBufferModule
 from .memory_bank_module import MemoryBankMixin
 
-__all__ = ["AnomalyModule", "BufferListMixin", "DynamicBufferModule", "MemoryBankMixin"]
+__all__ = ["AnomalyModule", "BufferListMixin", "DynamicBufferModule", "MemoryBankMixin", "ExportType"]
