+++ tests/integration/tools/test_gradio_entrypoint.py
@@ -12,7 +12,7 @@ from pathlib import Path
 import pytest
 
 from anomalib import TaskType
-from anomalib.deploy import OpenVINOInferencer, TorchInferencer, export_to_openvino, export_to_torch
+from anomalib.deploy import OpenVINOInferencer, TorchInferencer
 from anomalib.models import Padim
 
 sys.path.append("tools/inference")
@@ -46,8 +46,7 @@ class TestGradioInferenceEntrypoint:
         model = Padim.load_from_checkpoint(_ckpt_path)
 
         # export torch model
-        export_to_torch(
-            model=model,
+        model.to_torch(
             export_root=_ckpt_path.parent.parent,
             transform=transforms_config,
             task=TaskType.SEGMENTATION,
@@ -73,9 +72,8 @@ class TestGradioInferenceEntrypoint:
         model = Padim.load_from_checkpoint(_ckpt_path)
 
         # export OpenVINO model
-        export_to_openvino(
+        model.to_openvino(
             export_root=_ckpt_path.parent.parent,
-            model=model,
             input_size=(256, 256),
             transform=transforms_config,
             ov_args={},
