+++ tests/unit/deploy/test_inferencer.py
@@ -12,9 +12,9 @@ import torch
 
 from anomalib import TaskType
 from anomalib.data import MVTec
-from anomalib.deploy import ExportType, OpenVINOInferencer, TorchInferencer
+from anomalib.deploy import OpenVINOInferencer, TorchInferencer
 from anomalib.engine import Engine
-from anomalib.models import Padim
+from anomalib.models import ExportType, Padim
 
 
 class _MockImageLoader:
